requirements.txt
python-dotenv>=0.10
requests>=2.21
argparse>=1.4